Test: retryTx aspect 에 대한 테스트
- retry 정상 작동 여부 테스트
- hashTag 관련한 변경으로 인해 실패하는 테스트 수정
- aspect 시 annotation 이 제대로 들어가지 않는 문제 해결
- 주석 추가

Ref: #142

+++ src/main/java/com/studypals/domain/groupManage/dao/HashTagRepository.java
@@ -60,7 +60,7 @@ public interface HashTagRepository extends JpaRepository<HashTag, Long> {
             t.deletedAt = null
         WHERE t.tag = :tag
     """)
-    Long increaseUsedCount(String tag);
+    Integer increaseUsedCount(String tag);
 
     /**
      * usedCount 값을 원자적으로 증가시키는 메서드입니다. 단, 여러 tags 들에 대해 연산을 수행합니다.
